feat: Schema components
